fix issues with thread and add pop up

+++ src/main/java/action/upgrades/model/GroupData.java
@@ -3,8 +3,8 @@ package action.upgrades.model;
 public class GroupData {
     public String category;
     public String boost;
-    public int totalAmount = 0;
-    public int count = 0;
+    public long totalAmount = 0;
+    public long count = 0;
 
     public GroupData(String category, String boost) {
         this.category = category;
